+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ Released on: ???
 
 ## Features
 
+- Added Galatea water tracker which shows The Loch Emperor/Nessie catch statistics in Galatea [disabled by default].
 - Show Nessie's immunity flag in Sea creature HP overlay, while Nessie is running away.
 - Show player name who summoned a Reindrake, in the "Alert on ANY Reindrake" functionality.